+++ cmd/forklift/dev/plt/pallets.go
@@ -50,7 +50,7 @@ func processFullBaseArgs(
 	}
 	if opts.enableOverrides {
 		if caches.p, err = overlayPalletCacheOverrides(
-			caches.p.Underlay, c.StringSlice("plts"), plt,
+			caches.p.Underlay, c.StringSlice("plt"), plt,
 		); err != nil {
 			return nil, workspaceCaches{}, err
 		}
@@ -67,7 +67,7 @@ func processFullBaseArgs(
 	}
 	if opts.enableOverrides {
 		if caches.r, err = overlayRepoCacheOverrides(
-			caches.r, c.StringSlice("repos"), plt,
+			caches.r, c.StringSlice("repo"), plt,
 		); err != nil {
 			return nil, workspaceCaches{}, err
 		}
@@ -530,7 +530,12 @@ func lsPltFileAction(c *cli.Context) error {
 	if err != nil {
 		return nil
 	}
-	paths, err := fcli.ListPalletFiles(plt, c.Args().Get(1))
+	filter := c.Args().Get(1)
+	if filter == "" {
+		// Exclude hidden directories such as `.git`
+		filter = "{*,[^.]*/**}"
+	}
+	paths, err := fcli.ListPalletFiles(plt, filter)
 	if err != nil {
 		return err
 	}
@@ -577,4 +582,38 @@ func showPltFileAction(c *cli.Context) error {
 		return nil
 	}
 	return fcli.PrintFile(plt, c.Args().Get(1))
+}
+
+// ls-plt-feat
+
+func lsPltFeatAction(c *cli.Context) error {
+	plt, caches, err := processFullBaseArgs(c, processingOptions{
+		enableOverrides: true,
+	})
+	if err != nil {
+		return err
+	}
+
+	plt, err = fcli.GetRequiredPallet(plt, caches.p, c.Args().First())
+	if err != nil {
+		return nil
+	}
+	return fcli.PrintPalletFeatures(0, plt)
+}
+
+// show-plt-feat
+
+func showPltFeatAction(c *cli.Context) error {
+	plt, caches, err := processFullBaseArgs(c, processingOptions{
+		enableOverrides: true,
+	})
+	if err != nil {
+		return err
+	}
+
+	plt, err = fcli.GetRequiredPallet(plt, caches.p, c.Args().First())
+	if err != nil {
+		return nil
+	}
+	return fcli.PrintFeatureInfo(0, plt, caches.p, c.Args().Get(1))
 }