+++ cmd/forklift/plt/cli.go
@@ -220,6 +220,7 @@ func makeQuerySubcmds() []*cli.Command {
 		makeQueryFileSubcmds(category),
 		makeQueryPkgSubcmds(category),
 		makeQueryDeplSubcmds(category),
+		makeQueryFeatSubcmds(category),
 		[]*cli.Command{
 			{
 				Name:     "ls-dl",
@@ -271,6 +272,7 @@ func makeQueryReqSubcmds(category string) []*cli.Command {
 			},
 		},
 		makeQueryPltFileSubcmds(category),
+		makeQueryPltFeatSubcmds(category),
 		[]*cli.Command{
 			{
 				Name:     "ls-repo",
@@ -307,7 +309,7 @@ func makeQueryPltFileSubcmds(category string) []*cli.Command {
 			Category: category,
 			Usage: "Lists non-directory files in the specified pallet which the local pallet may " +
 				"import files from",
-			ArgsUsage: "[path_glob]",
+			ArgsUsage: "pallet_path [path_glob]",
 			Action:    lsPltFileAction,
 		},
 		{
@@ -316,7 +318,7 @@ func makeQueryPltFileSubcmds(category string) []*cli.Command {
 			Category: category,
 			Usage: "Prints the absolute filesystem path of the specified file in the specified " +
 				"pallet which the local pallet may import files from",
-			ArgsUsage: "file_path",
+			ArgsUsage: "pallet_path file_path",
 			Action:    locatePltFileAction,
 		},
 		{
@@ -325,8 +327,31 @@ func makeQueryPltFileSubcmds(category string) []*cli.Command {
 			Category: category,
 			Usage: "Prints the specified file in the specified pallet which the local pallet may " +
 				"import files from",
-			ArgsUsage: "file_path",
+			ArgsUsage: "pallet_path file_path",
 			Action:    showPltFileAction,
+		},
+	}
+}
+
+func makeQueryPltFeatSubcmds(category string) []*cli.Command {
+	return []*cli.Command{
+		{
+			Name:     "ls-plt-feat",
+			Aliases:  []string{"list-pallet-features"},
+			Category: category,
+			Usage: "Lists features flags exposed by the specified pallet which the local pallet may " +
+				"import files from",
+			ArgsUsage: "pallet_path",
+			Action:    lsPltFeatAction,
+		},
+		{
+			Name:     "show-plt-feat",
+			Aliases:  []string{"show-pallet-feature"},
+			Category: category,
+			Usage: "Prints the specified feature exposed by the specified pallet which the local " +
+				"pallet may import files from",
+			ArgsUsage: "pallet_path feature_name",
+			Action:    showPltFeatAction,
 		},
 	}
 }
@@ -436,6 +461,26 @@ func makeQueryDeplSubcmds(category string) []*cli.Command {
 					Usage: "Locates the package even if the specified deployment is disabled",
 				},
 			},
+		},
+	}
+}
+
+func makeQueryFeatSubcmds(category string) []*cli.Command {
+	return []*cli.Command{
+		{
+			Name:     "ls-feat",
+			Aliases:  []string{"list-features"},
+			Category: category,
+			Usage:    "Lists the feature flags exposed by the local pallet for other pallets to import",
+			Action:   lsFeatAction,
+		},
+		{
+			Name:      "show-feat",
+			Aliases:   []string{"show-feature"},
+			Category:  category,
+			Usage:     "Describes a feature exposed by the local pallet for other pallets to import",
+			ArgsUsage: "feature_name",
+			Action:    showFeatAction,
 		},
 	}
 }
@@ -686,8 +731,9 @@ func makeModifyDeplSubcmds( //nolint:funlen // this is already decomposed; it's 
 			Flags: slices.Concat(
 				[]cli.Flag{
 					&cli.StringSliceFlag{
-						Name:  "feature",
-						Usage: "Enable the specified feature flag in the package deployment",
+						Name:    "feat",
+						Aliases: []string{"feature", "features"},
+						Usage:   "Enable the specified feature in the package deployment",
 					},
 					&cli.BoolFlag{
 						Name:  "disabled",
@@ -726,7 +772,7 @@ func makeModifyDeplSubcmds( //nolint:funlen // this is already decomposed; it's 
 					&cli.BoolFlag{
 						Name: "force",
 						Usage: "Use the specified package path even if it cannot be resolved or makes the " +
-							"enabled feature flags invalid",
+							"enabled package features invalid",
 					},
 				},
 				modifyDeplBaseFlags,
@@ -749,7 +795,7 @@ func makeModifyDeplSubcmds( //nolint:funlen // this is already decomposed; it's 
 				[]cli.Flag{
 					&cli.BoolFlag{
 						Name: "force",
-						Usage: "Enable the specified feature flags even if they're not allowed by the  " +
+						Usage: "Enable the specified package features even if they're not allowed by the " +
 							"deployment's package",
 					},
 				},