+++ internal/app/forklift/pallets-models.go
@@ -177,9 +177,9 @@ type ResolvedImport struct {
 
 // An Import is an import group, a declaration of a group of files to import from a required pallet.
 type Import struct {
-	// Name is the name of the package file import.
+	// Name is the name of the file import group.
 	Name string
-	// Def is the file import definition for the file import.
+	// Def is the file import definition for the file import group.
 	Def ImportDef
 }
 
@@ -192,6 +192,9 @@ type ImportDef struct {
 	Modifiers []ImportModifier `yaml:"modifiers"`
 	// Disabled represents whether the import should be ignored.
 	Disabled bool `yaml:"disabled,omitempty"`
+	// Deprecated is a deprecation notice which, if specified as a non-empty string, causes warnings
+	// to be issued whenever the file import group is used via a feature flag.
+	Deprecated string `yaml:"deprecated,omitempty"`
 }
 
 // An ImportModifier defines an operation for transforming a set of files for importing into a
@@ -199,11 +202,14 @@ type ImportDef struct {
 type ImportModifier struct {
 	// Description is a short description of the import modifier to be shown to users.
 	Description string `yaml:"description,omitempty"`
-	// Type is either `add` (for adding one or more files to the set of files to import) or `remove`
-	// (for removing one or more files from the set of files to import)
+	// Type is either `add` (for adding one or more files to the set of files to import), `remove`
+	// (for removing one or more files from the set of files to import), `add-feature` (for adding
+	// files specified by a feature flag to the set of files to import), or `remove-feature` (for
+	// removing one or more files specified by a feature flag from the set of files to import).
 	Type string `yaml:"type,omitempty"`
 	// Source is the path in the required pallet of the file/directory to be imported, for an `add`
-	// modifier. If omitted, the source path will be inferred from the Target path.
+	// modifier; or the name of a feature flag, for an `add-feature` or `remove-feature` modifier. If
+	// omitted, the source path will be inferred from the Target path.
 	Source string `yaml:"source,omitempty"`
 	// Target is the path which the file/directory will be imported as, for an `add` modifier; or the
 	// path of the file/directory which will be removed from the set of files to import, for a
@@ -217,6 +223,18 @@ type ImportModifier struct {
 }
 
 const (
-	ImportModifierTypeAdd    = "add"
-	ImportModifierTypeRemove = "remove"
+	ImportModifierTypeAdd           = "add"
+	ImportModifierTypeRemove        = "remove"
+	ImportModifierTypeAddFeature    = "add-feature"
+	ImportModifierTypeRemoveFeature = "remove-feature"
+)
+
+// Features
+
+const (
+	// FeaturesDirName is the directory in a pallet containing declarations of file import groups
+	// which can be referenced by name in file import groups.
+	FeaturesDirName = "features"
+	// FeatureDefFileExt is the file extension for import group files.
+	FeatureDefFileExt = ".feature.yml"
 )